Move tests to src/__tests__ dir
