+++ package.json
@@ -4,8 +4,8 @@
   "description": "Mock Pusher.js in your JavaScript tests",
   "main": "index.js",
   "scripts": {
-    "test": "cross-env BABEL_ENV=commonjs mocha --compilers js:babel-register --recursive",
-    "cover": "cross-env BABEL_ENV=commonjs istanbul cover _mocha -- --compilers js:babel-register --recursive",
+    "test": "jest test --watch",
+    "cover": "jest test --coverage",
     "lint": "eslint src test",
     "clean": "rimraf lib"
   },
